Modify scripts to alert me when importing, renamed package and module to make the hierarchy clearer.
